+++ libraries/apollo-websocket-network-transport-incubating/README.md
@@ -1,4 +1,6 @@
 # Module apollo-websocket-network-transport-incubating
 
-An incubating network transport for websockets. See https://github.com/apollographql/apollo-kotlin/issues/5648 for more details
+An incubating network transport for websockets. It uses the `com.apollographql.apollo3.network.ws.incubating` package. The current plan is to drop the `.incubating` and replace the built-in `WebSocketNetworkTransport` for 4.0.0.
+
+See https://github.com/apollographql/apollo-kotlin/issues/5648 for more details
