+++ libraries/apollo-websocket-network-transport-incubating/src/commonMain/kotlin/com/apollographql/apollo3/network/ws/incubating/AppSyncWsProtocol.kt
@@ -1,5 +1,6 @@
 package com.apollographql.apollo3.network.ws.incubating
 
+import com.apollographql.apollo3.annotations.ApolloExperimental
 import com.apollographql.apollo3.api.ApolloRequest
 import com.apollographql.apollo3.api.NullableAnyAdapter
 import com.apollographql.apollo3.api.Operation
@@ -12,6 +13,10 @@ import com.apollographql.apollo3.api.json.writeAny
 import com.apollographql.apollo3.api.toJsonString
 import okio.Buffer
 
+/**
+ * A [WsProtocol] for https://docs.aws.amazon.com/appsync/latest/devguide/real-time-websocket-client.html
+ */
+@ApolloExperimental
 class AppSyncWsProtocol(
     val authorization: suspend () -> Any? = { null },
 ) : WsProtocol {
