+++ libraries/apollo-websocket-network-transport-incubating/src/commonMain/kotlin/com/apollographql/apollo3/network/ws/incubating/ServerMessage.kt
@@ -1,12 +1,22 @@
 package com.apollographql.apollo3.network.ws.incubating
 
+import com.apollographql.apollo3.annotations.ApolloExperimental
 import com.apollographql.apollo3.api.json.ApolloJsonElement
 
+/**
+ * A WebSocket [message](https://datatracker.ietf.org/doc/html/rfc6455#section-1.2) sent by the server
+ */
+@ApolloExperimental
 sealed interface ServerMessage
+@ApolloExperimental
 object ConnectionAckServerMessage : ServerMessage
+@ApolloExperimental
 object ConnectionKeepAliveServerMessage : ServerMessage
+@ApolloExperimental
 object PingServerMessage : ServerMessage
+@ApolloExperimental
 object PongServerMessage : ServerMessage
+@ApolloExperimental
 class ConnectionErrorServerMessage(val payload: ApolloJsonElement) : ServerMessage
 
 /**
@@ -15,12 +25,14 @@ class ConnectionErrorServerMessage(val payload: ApolloJsonElement) : ServerMessa
  * @param response, a GraphQL response, possibly containing errors.
  * @param complete, whether this is a terminal message for the given operation.
  */
+@ApolloExperimental
 class ResponseServerMessage(val id: String, val response: ApolloJsonElement, val complete: Boolean) : ServerMessage
 
 /**
  * The subscription completed normally
  * This is a terminal message for the given operation.
  */
+@ApolloExperimental
 class CompleteServerMessage(val id: String) : ServerMessage
 
 /**
@@ -29,9 +41,11 @@ class CompleteServerMessage(val id: String) : ServerMessage
  * @param payload additional information regarding the error. It may represent a GraphQL error
  * but it doesn't have to
  */
+@ApolloExperimental
 class OperationErrorServerMessage(val id: String, val payload: ApolloJsonElement, val terminal: Boolean) : ServerMessage
 
 /**
  * Special Server message that indicates a malformed message
  */
+@ApolloExperimental
 class ParseErrorServerMessage(val errorMessage: String) : ServerMessage
