+++ libraries/apollo-websocket-network-transport-incubating/src/commonMain/kotlin/com/apollographql/apollo3/network/ws/incubating/SubscriptionParser.kt
@@ -6,11 +6,17 @@ import com.apollographql.apollo3.api.ApolloResponse
 import com.apollographql.apollo3.api.Operation
 import com.apollographql.apollo3.api.json.ApolloJsonElement
 
+/**
+ * A [SubscriptionParser] transforms JSON responses contained in WebSocket messages into parsed [ApolloResponse]
+ */
 @ApolloExperimental
 interface SubscriptionParser<D : Operation.Data> {
-  fun parse(payload: ApolloJsonElement): ApolloResponse<D>?
+  fun parse(response: ApolloJsonElement): ApolloResponse<D>?
 }
 
+/**
+ * A factory for [SubscriptionParser]
+ */
 @ApolloExperimental
 interface SubscriptionParserFactory {
   fun <D: Operation.Data> createParser(request: ApolloRequest<D>): SubscriptionParser<D>
